+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 [0.1.1] - 2018-04-12
 --------------------
 - [x] Make base url-safe
-- [ ] Make constructor options optional #4
+- [x] Make constructor options optional #4
 
 
 [0.1.0] - 2018-04-02